experiment: refactored spotify handlers
