+++ src/core/spotify/ItemHandlers/AlbumHandler.ts
@@ -1,0 +1,13 @@
+import { AxiosInstance } from 'axios';
+import { ResultUrl } from '../../abstract/UrlHandler';
+import { getCachedSpotifyToken } from '../CachedSpotifyToken';
+import SpotifyApiWrapper from '../SpotifyApiWrapper';
+import { ItemHandler } from './ItemHandler.abstract';
+
+export class AlbumHandler extends ItemHandler {
+    public async handleItem(
+        item: string,
+    ): Promise<ResultUrl | ResultUrl[] | null> {
+        return SpotifyApiWrapper.getAlbum(item);
+    }
+}
