+++ src/core/spotify/ItemHandlers/ItemHandler.abstract.ts
@@ -4,6 +4,5 @@ import { ResultUrl } from '../../abstract/UrlHandler';
 export abstract class ItemHandler {
     public abstract handleItem(
         item: string,
-        instance: AxiosInstance,
     ): Promise<ResultUrl | ResultUrl[] | null>;
 }
