+++ src/core/spotify/SpotifyApiWrapper.ts
@@ -42,6 +42,67 @@ class SpotifyApiWrapper extends HttpClient {
             title: `${result.data.name} - ${result.data.artists[0].name}`,
         };
     }
+
+    async getPlaylist(itemId: string) {
+        const token = await getCachedSpotifyToken();
+
+        const response = await this.instance
+            .get(`/playlists/${itemId}/tracks`, {
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+                params: {
+                    fields: 'items(track(external_urls))',
+                },
+            })
+            .then((_res) => {
+                const items: {
+                    track: {
+                        external_urls: { spotify: string };
+                    };
+                }[] = _res.data.items;
+
+                const urls = items.map((value) => {
+                    return value.track.external_urls.spotify;
+                });
+
+                return urls;
+            });
+
+        return response.map((url) => {
+            return {
+                url,
+            };
+        });
+    }
+
+    async getAlbum(itemId: string) {
+        const token = await getCachedSpotifyToken();
+
+        const response = await this.instance
+            .get(`/albums/${itemId}/tracks`, {
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                },
+                params: {
+                    limit: 50,
+                },
+            })
+            .then((res) => {
+                const items: { external_urls: { spotify: string } }[] =
+                    res.data.items;
+
+                return items.map((item) => {
+                    return item.external_urls.spotify;
+                });
+            });
+
+        return response.map((url) => {
+            return {
+                url,
+            };
+        });
+    }
 }
 
 export default new SpotifyApiWrapper();
